feat: added `MoveGen` struct that generates moves as an Iterator instead of just generating them all at once. Still a work-in-progress, but it passes all perft suites, so it functions properly
