+++ chessie/types/src/utils.rs
@@ -11,7 +11,9 @@ pub const FEN_STARTPOS: &str = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQ
 /// A popular FEN string for debugging move generation.
 pub const FEN_KIWIPETE: &str = "r3k2r/p1ppqpb1/bn2pnp1/3PN3/1p2P3/2N2Q1p/PPPBBPPP/R3K2R w KQkq -";
 
-/// <https://www.chessprogramming.org/Chess_Position#cite_note-4>
+/// Maximum possible number of moves in a given chess position.
+///
+/// Found [here](<https://www.chessprogramming.org/Chess_Position#cite_note-4>)
 pub const MAX_NUM_MOVES: usize = 218;
 
 /// Number of possible combinations of castling rights.
